script_generator/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-generativeai==0.8.3
pydantic==2.6.4
